pyproject.toml
[project]
name = "demo-temporal"
version = "0.1.0"
description = "Temporal CRM demo — workflows, Saga compensations, multi-replica workers"
requires-python = ">=3.13,<3.14"
dependencies = [
    "temporalio[pydantic]>=1.9.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.20",
    "asyncpg>=0.30.0",
    "pydantic>=2.10.0",
    "structlog>=25.0.0",
    "python-ulid>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.11.0",
    "pyrefly>=0.61.0",
    "httpx>=0.28.1",
]

[tool.uv]
package = false

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "RUF"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = [
    "activities",
    "api",
    "common",
    "contracts",
    "gateways",
    "observability",
    "repo",
    "worker",
    "workflows",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyrefly]
project_includes = ["src"]
python_version = "3.13"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
python_files = ["test_*.py"]
pythonpath = ["src"]
